INT-17366: tool_ally: Cache annotation_maps
Without this, every time a user hits a page, annotation maps are calculated for
the course - these maps are the same for all users and require many db reads to
generate, especially for large courses.

Caching these for a few minutes helps relieve pressure on the db during periods
a single course is very busy (quiz or assignment open/close times for example).

Unfortunately, adding/editing blocks doesn't trigger any events that we can
hook into to purge the annotation map cache.

What we _can_ do however - is limit the annotation map reload when the cache is
available to only refreshing data for blocks. Currently that is just the html
block, which loads all instance in one query for the course, which isn't too
bad.